Fixed the particle effects on collision with asteroids
I found out that the size of the backgrounds and where I had the spawnner was causing my particle effects to be overlapped. I scaled down the size of the backgrounds to 1.5 and changed the starting X Position on space1 to 5 and space2 to 25. This allowed me to easy calculate the endX and startX for the repeating BG script to be consistent.